Moved getAddressText to helpers in address-entry so that it can be reused in contact-entry, as suggested by code climate
