+++ src/internal/modules/address-entry/lib/helpers.js
@@ -1,7 +1,8 @@
 const sessionHelpers = require('shared/lib/session-helpers');
 const services = require('../../../lib/connectors/services');
 const queryString = require('querystring');
-
+const { compact } = require('lodash');
+const titleCase = require('title-case');
 const SESSION_KEY = 'newAddressFlow';
 
 const getSession = request => request.yar.get(SESSION_KEY);
@@ -56,6 +57,14 @@ const getPageCaption = request => {
   if (session && session.licenceNumber) return { caption: `Licence ${session.licenceNumber}` };
 };
 
+const getAddressText = address => {
+  const { addressLine1, addressLine2, addressLine3, addressLine4 } = address;
+  const addressLines = compact([addressLine1, addressLine2, addressLine3, addressLine4])
+    .map(line => titleCase(line))
+    .join(' ');
+  return `${addressLines}, ${titleCase(address.town)}, ${address.postcode}`;
+};
+
 exports.SESSION_KEY = SESSION_KEY;
 
 exports.getRedirectPath = getRedirectPath;
@@ -65,3 +74,4 @@ exports.getPostcodeUrlParams = getPostcodeUrlParams;
 exports.getPostcodeUrl = getPostcodeUrl;
 exports.getManualAddressEntryBackLink = getManualAddressEntryBackLink;
 exports.saveReferenceData = saveReferenceData;
+exports.getAddressText = getAddressText;
