+++ src/internal/modules/contact-entry/forms/company-search-select-address.js
@@ -1,8 +1,7 @@
 const { formFactory, fields } = require('shared/lib/forms');
 const Joi = require('@hapi/joi');
-
 const { uniqBy } = require('lodash');
-const { getAddressText } = require('../helpers');
+const { getAddressText } = require('../../address-entry/lib/helpers');
 
 const getAddressChoices = addresses => {
   const choices = uniqBy(addresses, entity => [entity.postcode].join()).map(record => {
